ImmutableJS in Post and PostList reducers (#78)
* refactor client, common, server

* bump node version to 6.2.1

* missing comma in package.json

* update travis

* not using RHL anymore

* fix text decoration on the about page

* clean up webpack, swap `indexRoute`

This restore HMR. It also swaps the `indexRoute` of the application.
The About page is now the primary view. PostList is now at `/posts`.

* Rename About -> Home, remove PostList reducer

* clean up styles

* drop API middleware,

Middleware is pretty opinionated, in order to make things work I would
have needed to use a custom solution. This new code is agnostic with
respect to whether you use `isomorphic-fetch` or `axios` or whatevs.

* delete middleware docs

* add demo selectors for reducers (#72)

* add demo selectors for reducers

* fix tests

* Immutable added as a dependency

* Post and PostList reducer now using Immutable.js

- tests updated
- chai and chai-http moved to devDeps