Make React Razzle project easily deployable
I noticed that Razzle seems to be missing a deployment story. Deploying with `now` by default doesn't work because the default `start` command is for dev command only.

So I added a couple of lines to `package.json` and now (sic) it works. 

```
"now-start": "npm run start:prod",
"deploy": "now -e NODE_ENV=production"
```

I think it could be useful seeing the project live on a url from the get-go.