react-loadable example: fix bundle loading order
in the react-loadable example the bundles are currently not loaded in
the correct order: the "chunks" are loaded before the main bundle is
loaded. This does not work because the webpack bundle functions required
to load the chunks are not yet loaded, leading to this error:
"0.51b7694d.chunk.js:1 Uncaught ReferenceError: webpackJsonp is not
defined"

This commit changes server.js to follow the react-loadable docs more
closely. It removes the defer attribute from the main bundle, so that it
is loaded before the chunks.
Also uses Loadable.preloadReady() to defer hydration until all chunks
are loaded, as described in the react-loadable docs.