react-loadable example: fix chunk path in dev
when running the dev-server, the preloaded chunks for react-loadable
must be loaded from the dev-server running at port 3001(by default).
Before this commit they were loaded with relative path - this worked in
production, but in dev mode the index.html was served instead, resulting
in "Uncaught SyntaxError: Unexpected token <" errors when parsing the
chunks.