Use Yarn Workspaces
Removes manual linking and direct bootstrapping with Lerna in favor of Yarn Workspaces. See facebook/create-react-app#3755.

Before using Yarn Workspaces, tests were not passing because of an error similar to babel/babel#7110; it seems it was related to conflicting babel-core (see https://github.com/babel/babel/issues/7110#issuecomment-354583771) since we are using Jest which depends on babel-core 6 and needs babel-core 7-bridge to work.

Issues disappear with Yarn Workspaces and tests are passing.