Updating includes to be future-safe, fixing the suggestions in warnings.
This is allowing to remove the following warnings and making it future safe:
```
Warning: Please use `require("react-router-dom").Route` instead of `require("react-router-dom/Route")`. Support for the latter will be removed in the next major release.
Warning: Please use `require("react-router-dom").Switch` instead of `require("react-router-dom/Switch")`. Support for the latter will be removed in the next major release.```