Add documentation for non-relative module imports.
Non-relative module support isn't currently documented except for a comment in #1112. I wasn't aware that you could also set module base directories in package.json -> jest, so it took me a lot of back and forth before I finally figured it out. I think rather than trying to automate the jest configuration based on the `baseUrl` setting, an update to the documentation is a more appropriate solution. This addresses #1466.