fix(razzle): Exit start command on unhandled rejections
Currently, if code inside the razzle.config.js (e.g.
modifyWebpackConfig) throws an error, the process just quits without
displaying any error message.

Long-term it might be better to catch this error a bit closer to the
code that is causing it (e.g. "modifyWebpackConfig raised an error"),
but for now this aligns the handling to what the build and test command
are already doing.